Clear all warning errors + missing depencies
