+++ src/config/providers/APIProvider/client.ts
@@ -1,29 +1,8 @@
-import { ApolloClient, FieldPolicy, InMemoryCache } from "@apollo/client";
-import { offsetLimitPagination, Reference } from "@apollo/client/utilities";
+import { ApolloClient, InMemoryCache } from "@apollo/client";
 import { RestLink } from "apollo-link-rest";
-import { PagedResultsList } from "types";
+import { mangadexOffsetLimitPagination } from "./utilities";
 
 const restLink = new RestLink({ uri: "http://localhost:3001" });
-
-type KeyArgs = FieldPolicy<any>["keyArgs"];
-
-function mangadexOffsetLimitPagination<T extends Reference>(
-  keyArgs: KeyArgs = false
-): FieldPolicy<PagedResultsList<T>> {
-  return {
-    keyArgs,
-    merge(existing, incoming) {
-      if (!existing) {
-        return incoming;
-      }
-
-      return {
-        ...incoming,
-        results: [...existing.results, ...incoming.results],
-      };
-    },
-  };
-}
 
 const client = new ApolloClient({
   cache: new InMemoryCache({
